refactor(app): split directive handlers
